+++ AppleModels/update_generated.py
@@ -138,11 +138,11 @@ def export_db_macinfolib(db, path, year=0):
       if max(info['AppleModelYear']) < year:
         continue
 
-      br_model = info.get('AppleModelId')
-      if br_model:
-        br_model = '"{}AP"'.format(br_model.upper())
+      sb_model = info.get('AppleModelId')
+      if sb_model:
+        sb_model = '"{}"'.format(sb_model)
       else:
-        br_model = 'NULL'
+        sb_model = 'NULL'
 
       print(' {\n'
         '  .SystemProductName = "%s",\n'
@@ -167,7 +167,7 @@ def export_db_macinfolib(db, path, year=0):
         '  .ChassisAssetTag = "%s",\n'
         '  .FirmwareFeatures = 0x%XULL,\n'
         '  .FirmwareFeaturesMask = 0x%XULL,\n'
-        '  .BridgeModel = %s,\n'
+        '  .SecureBootModel = %s,\n'
         ' },' % (
           info['SystemProductName'],
           info['BoardProduct'][0] if isinstance(info['BoardProduct'], list) else info['BoardProduct'],
@@ -191,7 +191,7 @@ def export_db_macinfolib(db, path, year=0):
           info['ChassisAssetTag'],
           info.get('ExtendedFirmwareFeatures', info.get('FirmwareFeatures', 0)),
           info.get('ExtendedFirmwareFeaturesMask', info.get('FirmwareFeaturesMask', 0)),
-          br_model
+          sb_model
         ), file=fh)
 
     print('};', file=fh)