+++ Library/OcAppleImg4Lib/OcAppleImg4Lib.c
@@ -42,7 +42,6 @@ typedef struct OC_SB_MODEL_DESC_ {
 
 STATIC CHAR8 mCryptoDigestMethod[16] = "sha2-384";
 STATIC DERImg4Environment mEnvInfo;
-STATIC CONST CHAR8 *mModelDefault = "x86legacy";
 ///
 /// List of model mapping to board identifiers.
 /// Alphabetically sorted (!), for release order refer to the documentation.
@@ -64,7 +63,7 @@ STATIC OC_SB_MODEL_DESC mModelInformation[] = {
   { "j230k",     0x3F }, ///< MacBookAir9,1
   { "j680",      0x0B }, ///< MacBookPro15,1
   { "j780",      0x07 }, ///< MacBookPro15,3
-  { "x86legacy", 0xF0 }, ///< Generic x86 from Big Sur
+  { OC_SB_MODEL_LEGACY, 0xF0 }, ///< Generic x86 from Big Sur
 };
 
 STATIC BOOLEAN mHasDigestOverride;
@@ -159,10 +158,6 @@ OcAppleImg4GetHardwareModel (
   )
 {
   OC_SB_MODEL_DESC  *ModelInfo;
-
-  if (AsciiStrCmp (ModelRequest, OC_SB_MODEL_DEFAULT) == 0 || ModelRequest[0] == '\0') {
-    ModelRequest = mModelDefault;
-  }
 
   ModelInfo = InternalGetModelInfo (ModelRequest);
   if (ModelInfo != NULL) {
@@ -362,7 +357,7 @@ OcAppleImg4BootstrapValues (
   mEnvInfo.securityMode               = 1;
   mEnvInfo.effectiveProductionStatus  = TRUE;
   mEnvInfo.effectiveSecurityMode      = 1;
-  mEnvInfo.internalUseOnlyUnit        = AsciiStrCmp (Model, "x86legacy") == 0;
+  mEnvInfo.internalUseOnlyUnit        = AsciiStrCmp (Model, OC_SB_MODEL_LEGACY) == 0;
   mEnvInfo.xugs                       = 1;
   mEnvInfo.allowMixNMatch             = FALSE;