+++ Library/OcMacInfoLib/AutoGenerated.c
@@ -24,7 +24,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook10,1",
@@ -49,7 +49,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FC0FE13FULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook2,1",
@@ -74,7 +74,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook3,1",
@@ -99,7 +99,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001407ULL,
   .FirmwareFeaturesMask = 0xC0001FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook4,1",
@@ -124,7 +124,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0003FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook5,1",
@@ -149,7 +149,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0007FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook5,2",
@@ -174,7 +174,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0007FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook6,1",
@@ -199,7 +199,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xFC0FE13FULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook7,1",
@@ -224,7 +224,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xFC0FE13FULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook8,1",
@@ -249,7 +249,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Chassis Board Asset Tag#",
   .FirmwareFeatures = 0xFC0FE13FULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBook9,1",
@@ -274,7 +274,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FC0FE13FULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir1,1",
@@ -299,7 +299,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0003FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir2,1",
@@ -324,7 +324,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0007FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir3,1",
@@ -349,7 +349,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xD00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir3,2",
@@ -374,7 +374,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xD00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir4,1",
@@ -399,7 +399,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xD00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir4,2",
@@ -424,7 +424,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xD00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir5,1",
@@ -449,7 +449,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir5,2",
@@ -474,7 +474,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir6,1",
@@ -499,7 +499,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00FE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir6,2",
@@ -524,7 +524,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00FE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir7,1",
@@ -549,7 +549,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Chassis Board Asset Tag#",
   .FirmwareFeatures = 0x8FF0FF576ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir7,2",
@@ -574,7 +574,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Chassis Board Asset Tag#",
   .FirmwareFeatures = 0x8FF0FF576ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookAir8,1",
@@ -599,7 +599,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF42EULL,
   .FirmwareFeaturesMask = 0x8FF9FFF3FULL,
-  .BridgeModel = "J140KAP",
+  .SecureBootModel = "J140K",
  },
  {
   .SystemProductName = "MacBookAir8,2",
@@ -624,7 +624,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF42EULL,
   .FirmwareFeaturesMask = 0x8FF9FFF3FULL,
-  .BridgeModel = "J140AAP",
+  .SecureBootModel = "J140A",
  },
  {
   .SystemProductName = "MacBookAir9,1",
@@ -649,7 +649,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FFAFF06EULL,
   .FirmwareFeaturesMask = 0x8FFFFFF7FULL,
-  .BridgeModel = "J230KAP",
+  .SecureBootModel = "J230K",
  },
  {
   .SystemProductName = "MacBookPro1,1",
@@ -674,7 +674,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro1,2",
@@ -699,7 +699,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro10,1",
@@ -724,7 +724,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro10,2",
@@ -749,7 +749,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro11,1",
@@ -774,7 +774,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xEB0FF577ULL,
   .FirmwareFeaturesMask = 0xFF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro11,2",
@@ -799,7 +799,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xEB0FF577ULL,
   .FirmwareFeaturesMask = 0xFF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro11,3",
@@ -824,7 +824,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xEB0FF577ULL,
   .FirmwareFeaturesMask = 0xFF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro11,4",
@@ -849,7 +849,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xEB0FF577ULL,
   .FirmwareFeaturesMask = 0xFF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro11,5",
@@ -874,7 +874,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xEB0FF577ULL,
   .FirmwareFeaturesMask = 0xFF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro12,1",
@@ -899,7 +899,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD0FF576ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro13,1",
@@ -924,7 +924,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FC0FE137ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro13,2",
@@ -949,7 +949,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FC0FE137ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro13,3",
@@ -974,7 +974,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FC0FE137ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro14,1",
@@ -999,7 +999,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FF0FF57EULL,
   .FirmwareFeaturesMask = 0x8FF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro14,2",
@@ -1024,7 +1024,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FF0FF57EULL,
   .FirmwareFeaturesMask = 0x8FF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro14,3",
@@ -1049,7 +1049,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FF0FF57EULL,
   .FirmwareFeaturesMask = 0x8FF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro15,1",
@@ -1074,7 +1074,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF426ULL,
   .FirmwareFeaturesMask = 0x8FF9FFF3FULL,
-  .BridgeModel = "J680AP",
+  .SecureBootModel = "J680",
  },
  {
   .SystemProductName = "MacBookPro15,2",
@@ -1099,7 +1099,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF426ULL,
   .FirmwareFeaturesMask = 0x8FF9FFF3FULL,
-  .BridgeModel = "J132AP",
+  .SecureBootModel = "J132",
  },
  {
   .SystemProductName = "MacBookPro15,3",
@@ -1124,7 +1124,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF426ULL,
   .FirmwareFeaturesMask = 0x8FF9FFF3FULL,
-  .BridgeModel = "J780AP",
+  .SecureBootModel = "J780",
  },
  {
   .SystemProductName = "MacBookPro15,4",
@@ -1149,7 +1149,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF426ULL,
   .FirmwareFeaturesMask = 0x8FF9FFF3FULL,
-  .BridgeModel = "J213AP",
+  .SecureBootModel = "J213",
  },
  {
   .SystemProductName = "MacBookPro16,1",
@@ -1174,7 +1174,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FDAFF066ULL,
   .FirmwareFeaturesMask = 0x8FFFFFF7FULL,
-  .BridgeModel = "J152FAP",
+  .SecureBootModel = "J152F",
  },
  {
   .SystemProductName = "MacBookPro16,2",
@@ -1199,7 +1199,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FFAFF06EULL,
   .FirmwareFeaturesMask = 0x8FFFFFF7FULL,
-  .BridgeModel = "J214KAP",
+  .SecureBootModel = "J214K",
  },
  {
   .SystemProductName = "MacBookPro16,3",
@@ -1224,7 +1224,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FDAFF066ULL,
   .FirmwareFeaturesMask = 0x8FFFFFF7FULL,
-  .BridgeModel = "J223AP",
+  .SecureBootModel = "J223",
  },
  {
   .SystemProductName = "MacBookPro16,4",
@@ -1249,7 +1249,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FDAFF066ULL,
   .FirmwareFeaturesMask = 0x8FFFFFF7FULL,
-  .BridgeModel = "J215AP",
+  .SecureBootModel = "J215",
  },
  {
   .SystemProductName = "MacBookPro2,1",
@@ -1274,7 +1274,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro2,2",
@@ -1299,7 +1299,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro3,1",
@@ -1324,7 +1324,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001407ULL,
   .FirmwareFeaturesMask = 0xC0001FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro4,1",
@@ -1349,7 +1349,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0003FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro5,1",
@@ -1374,7 +1374,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0007FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro5,2",
@@ -1399,7 +1399,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0007FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro5,3",
@@ -1424,7 +1424,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0007FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro5,4",
@@ -1449,7 +1449,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0007FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro5,5",
@@ -1474,7 +1474,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0007FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro6,1",
@@ -1499,7 +1499,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro6,2",
@@ -1524,7 +1524,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro7,1",
@@ -1549,7 +1549,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro8,1",
@@ -1574,7 +1574,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro8,2",
@@ -1599,7 +1599,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro8,3",
@@ -1624,7 +1624,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro9,1",
@@ -1649,7 +1649,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacBookPro9,2",
@@ -1674,7 +1674,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC10DF577ULL,
   .FirmwareFeaturesMask = 0xFF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacPro1,1",
@@ -1699,7 +1699,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0x80000015ULL,
   .FirmwareFeaturesMask = 0x800003FFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacPro2,1",
@@ -1724,7 +1724,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0000015ULL,
   .FirmwareFeaturesMask = 0xC0000FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacPro3,1",
@@ -1749,7 +1749,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Pro-Enclosure",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0003FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacPro4,1",
@@ -1774,7 +1774,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Pro-Enclosure",
   .FirmwareFeatures = 0xE001F537ULL,
   .FirmwareFeaturesMask = 0xC003FF37ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacPro5,1",
@@ -1799,7 +1799,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE80FE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacPro6,1",
@@ -1824,7 +1824,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8E90FF576ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "MacPro7,1",
@@ -1849,7 +1849,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FDAFF066ULL,
   .FirmwareFeaturesMask = 0x8FFFFFF7FULL,
-  .BridgeModel = "J160AP",
+  .SecureBootModel = "J160",
  },
  {
   .SystemProductName = "Macmini1,1",
@@ -1874,7 +1874,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini2,1",
@@ -1899,7 +1899,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini3,1",
@@ -1924,7 +1924,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0007FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini4,1",
@@ -1949,7 +1949,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC00C9423ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini5,1",
@@ -1974,7 +1974,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xD00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini5,2",
@@ -1999,7 +1999,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xD00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini5,3",
@@ -2024,7 +2024,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xD00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini6,1",
@@ -2049,7 +2049,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini6,2",
@@ -2074,7 +2074,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini7,1",
@@ -2099,7 +2099,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8E00DE137ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Macmini8,1",
@@ -2124,7 +2124,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF466ULL,
   .FirmwareFeaturesMask = 0x8FFDFFF7FULL,
-  .BridgeModel = "J174AP",
+  .SecureBootModel = "J174",
  },
  {
   .SystemProductName = "Xserve1,1",
@@ -2149,7 +2149,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Xserve2,1",
@@ -2174,7 +2174,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "Xserve3,1",
@@ -2199,7 +2199,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Xserve",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac10,1",
@@ -2224,7 +2224,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac11,1",
@@ -2249,7 +2249,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac11,2",
@@ -2274,7 +2274,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xC00C9423ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac11,3",
@@ -2299,7 +2299,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac12,1",
@@ -2324,7 +2324,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac12,2",
@@ -2349,7 +2349,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac13,1",
@@ -2374,7 +2374,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac13,2",
@@ -2399,7 +2399,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac13,3",
@@ -2424,7 +2424,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00DE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac14,1",
@@ -2449,7 +2449,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xFB0FF577ULL,
   .FirmwareFeaturesMask = 0xFF1FFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac14,2",
@@ -2474,7 +2474,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00FE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac14,3",
@@ -2499,7 +2499,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xE00FE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac14,4",
@@ -2524,7 +2524,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xF00FE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac15,1",
@@ -2549,7 +2549,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0xF80FE137ULL,
   .FirmwareFeaturesMask = 0xFF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac16,1",
@@ -2574,7 +2574,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FC0FE137ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac16,2",
@@ -2599,7 +2599,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FC0FE137ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac17,1",
@@ -2624,7 +2624,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FC0FE137ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac18,1",
@@ -2649,7 +2649,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD0FF576ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac18,2",
@@ -2674,7 +2674,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD0FF576ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac18,3",
@@ -2699,7 +2699,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD0FF576ULL,
   .FirmwareFeaturesMask = 0x8FF1FFF3FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac19,1",
@@ -2724,7 +2724,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF576ULL,
   .FirmwareFeaturesMask = 0x8FFDFFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac19,2",
@@ -2749,7 +2749,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF576ULL,
   .FirmwareFeaturesMask = 0x8FFDFFF7FULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac20,1",
@@ -2774,7 +2774,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF576ULL,
   .FirmwareFeaturesMask = 0x8FFDFFF7FULL,
-  .BridgeModel = "J185AP",
+  .SecureBootModel = "J185",
  },
  {
   .SystemProductName = "iMac20,2",
@@ -2799,7 +2799,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF576ULL,
   .FirmwareFeaturesMask = 0x8FFDFFF7FULL,
-  .BridgeModel = "J185FAP",
+  .SecureBootModel = "J185F",
  },
  {
   .SystemProductName = "iMac4,1",
@@ -2824,7 +2824,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac4,2",
@@ -2849,7 +2849,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac5,1",
@@ -2874,7 +2874,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac5,2",
@@ -2899,7 +2899,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac6,1",
@@ -2924,7 +2924,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag",
   .FirmwareFeatures = 0x0ULL,
   .FirmwareFeaturesMask = 0x0ULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac7,1",
@@ -2949,7 +2949,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001407ULL,
   .FirmwareFeaturesMask = 0xC0001FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac8,1",
@@ -2974,7 +2974,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0003FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMac9,1",
@@ -2999,7 +2999,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "Asset Tag#",
   .FirmwareFeatures = 0xC0001403ULL,
   .FirmwareFeaturesMask = 0xC0003FFFULL,
-  .BridgeModel = NULL,
+  .SecureBootModel = NULL,
  },
  {
   .SystemProductName = "iMacPro1,1",
@@ -3024,7 +3024,7 @@ CONST MAC_INFO_INTERNAL_ENTRY gMacInfoModels[] = {
   .ChassisAssetTag = "",
   .FirmwareFeatures = 0x8FD8FF53EULL,
   .FirmwareFeaturesMask = 0x8FF9FFF3FULL,
-  .BridgeModel = "J137AP",
+  .SecureBootModel = "J137",
  },
 };
 CONST UINTN gMacInfoModelCount = ARRAY_SIZE (gMacInfoModels);